+++ .husky/pre-commit
@@ -1,5 +1,6 @@
 #!/usr/bin/env sh
 . "$(dirname -- "$0")/_/husky.sh"
 
+yarn build
 yarn test
 npx lint-staged
