+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "@mashup-node/content-type",
   "private": false,
-  "version": "0.0.2",
+  "version": "0.0.4",
   "author": {
     "email": "mashup.node@gmail.com",
     "name": "MashUp Node"
